# src/auth/auth.controller.ts
import {
  Controller,
  Post,
  Get,
  Body,
  Res,
  Req,
  UseGuards,
  HttpCode,
  HttpStatus,
} from "@nestjs/common";
import {
  ApiTags,
  ApiOperation,
  ApiResponse,
  ApiBody,
  ApiCookieAuth,
  ApiBearerAuth,
  ApiProperty,
} from "@nestjs/swagger";
import { Response, Request } from "express";
import { User } from "@prisma/client";
import { IsEmail, IsString } from "class-validator";
import { AuthService } from "./auth.service";
import { JwtAuthGuard } from "./jwt-auth.guard";
import { RefreshTokenGuard } from "./refresh.guard";

export interface RequestWithUser extends Request {
  user: any;
}

class LoginDto {
  @ApiProperty({ example: 'admin@example.com' })
  @IsEmail()
  email: string;

  @ApiProperty({ example: 'admin123' })
  @IsString()
  password: string;
}

class LoginResponse {
  @ApiProperty({ example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...' })
  accessToken: string;

  @ApiProperty({ example: 900 })
  expiresIn: number;
}

@ApiTags("auth")
@Controller("auth")
export class AuthController {
  constructor(private readonly authService: AuthService) {}

  /**
   * Login: validate credentials, issue access & refresh tokens.
   * Refresh token is set as HttpOnly cookie.
   */
  @Post("login")
  @HttpCode(HttpStatus.OK)
  @ApiOperation({ summary: "Login user" })
  @ApiBody({ type: LoginDto })
  @ApiResponse({
    status: 200,
    description: "Login successful",
    type: LoginResponse,
  })
  @ApiResponse({ status: 401, description: "Unauthorized" })
  async login(
    @Body() loginDto: LoginDto,
    @Res() res: Response,
    @Req() req: Request
  ) {
    const deviceInfo = req.get("user-agent") || "unknown";
    const result = await this.authService.signInByEmail(
      loginDto.email,
      loginDto.password,
      deviceInfo
    );
    // result: { accessToken, expiresIn, accessJti, refreshToken, refreshJti, refreshExpiresAt, user }

    const ttlSeconds =
      Number(process.env.REFRESH_TOKEN_TTL) || 14 * 24 * 60 * 60;
    const secureCookie = process.env.NODE_ENV === "production";

    res.cookie("refresh_token", result.refreshToken, {
      httpOnly: true,
      secure: secureCookie,
      sameSite: "lax",
      maxAge: ttlSeconds * 1000,
      path: "/",
    });

    return res.json({
      accessToken: result.accessToken,
      expiresIn: result.expiresIn,
    });
  }

  /**
   * Refresh: Guard validates the refresh token (from cookie or header),
   * attaches token record to request. We rotate the refresh token and issue new access.
   */
  @Post("refresh")
  @UseGuards(RefreshTokenGuard)
  @HttpCode(HttpStatus.OK)
  @ApiOperation({ summary: "Refresh access token" })
  @ApiCookieAuth("refresh_token")
  @ApiResponse({
    status: 200,
    description: "Token refreshed",
    type: LoginResponse,
  })
  @ApiResponse({ status: 401, description: "Unauthorized" })
  async refresh(@Req() req: RequestWithUser, @Res() res: Response) {
    // Guard set these:
    const tokenRecord = (req as any).refreshTokenRecord;
    const user = tokenRecord.user;
    const oldJti = tokenRecord.jti;

    // rotate: mark old as revoked and create a new refresh token
    const { token: newRefreshToken } =
      await this.authService.rotateRefreshToken(
        oldJti,
        user.id,
        req.get("user-agent") || "unknown"
      );

    // create new access token
    const { accessToken, expiresIn } =
      await this.authService.createAccessToken(user);

    const ttlSeconds =
      Number(process.env.REFRESH_TOKEN_TTL) || 14 * 24 * 60 * 60;
    const secureCookie = process.env.NODE_ENV === "production";

    res.cookie("refresh_token", newRefreshToken, {
      httpOnly: true,
      secure: secureCookie,
      sameSite: "lax",
      maxAge: ttlSeconds * 1000,
      path: "/",
    });

    return res.json({
      accessToken,
      expiresIn,
    });
  }

  /**
   * Logout: revoke current refresh token (if any) and clear cookie.
   */
  @Post("logout")
  @HttpCode(HttpStatus.OK)
  @ApiOperation({ summary: "Logout user" })
  @ApiCookieAuth("refresh_token")
  @ApiResponse({ status: 200, description: "Logged out" })
  async logout(@Req() req: Request, @Res() res: Response) {
    const raw = req.cookies && req.cookies["refresh_token"];
    if (raw) {
      const tokenRecord = await this.authService.validateRefreshToken(raw);
      if (tokenRecord) {
        await this.authService.revokeRefreshTokenByJti(tokenRecord.jti);
      }
    }

    res.clearCookie("refresh_token", { path: "/" });
    return res.json({ message: "Logged out" });
  }

  /**
   * Get current user info (protected by access token)
   */
  @Get("me")
  @UseGuards(JwtAuthGuard)
  @ApiOperation({ summary: "Get current user info" })
  @ApiResponse({
    status: 200,
    description: "User info",
    schema: {
      type: 'object',
      properties: {
        id: { type: 'string', example: 'user_id_here' },
        email: { type: 'string', example: 'user@example.com' },
        name: { type: 'string', example: 'John Doe' },
      },
    },
  })
  @ApiResponse({ status: 401, description: "Unauthorized" })
  async me(@Req() req: RequestWithUser) {
    const { id, email, name } = req.user;
    return {
      id,
      email,
      name,
    };
  }
}